Allow (and force) user to insert '$' as '$$'
Before:     There was no way to insert '$' within words.
After:      '$$foo' %{} -> '$foo'

NOTE: I decided it was easier to drop allowing stand-alone '$' with
non-alphanumerical character following, i.e:

Example:
    p = \f print(pcall(f))
    p(\ '${}'   % {})
    p(\ '${'    % {})
    p(\ '$'     % {})
    p(\ '$$foo' % {})
    p(\ '$$foo' % {foo="bar"})

Before:
    true	${}
    true	${
    true	$
    true	$foo
    true	$bar

After:
    false	./src/madl_gutil.mad:431: Undefined key: "${}"
    false	./src/madl_gutil.mad:431: Undefined key: "${"
    false	./src/madl_gutil.mad:431: Undefined key: "$"
    true	$foo
    true	$foo

In fact, the previous behaviour could also be emulated using additional
logic in the `interp_fun`, but I don't think it's very useful or even
desirable at all. It's better to later add support for this if needed,
rather than allowing it from the start and having to live with it.