Error on undefined keys
Interpolating `nil` is almost always an error and should `error()`!

Example:
    p = \f print(pcall(f))
    p(\ "$1"     % {})
    p(\ "$foo"   % {})
    p(\ "$foo}"  % {})

Before:
    true	1
    true	foo
    true	foo}}

After:
    false	./src/madl_gutil.mad:430: Undefined key: "$1"
    false	./src/madl_gutil.mad:430: Undefined key: "$foo"
    false	./src/madl_gutil.mad:430: Undefined key: "$foo}"

The previous behaviour of interpolating missing values as the key name
itself is very undesirable and allows easy-to-miss errors.

If someone wants to allow interpolating `nil`, there can be an option or
they can simply use a metatable `{__index=\ "nil"}`.